KidneX/requirements.txt
numpy==1.23.5
streamlit==1.22.0
scikit-learn==0.23.0
Pillow==8.3.2
tensorflow==2.12.0
keras==2.12.0
opencv_python==4.6.0.66